+++ MatrixMultiplikation_C++/MatrixMultiplication/Matrix.h
@@ -6,7 +6,7 @@ namespace Matrix {
 	public:
 		const static int n = 1440;
 		const static int m = 1440;
-		long *matrix;
+		float *matrix;
 		void createRandomMatrix();
 		void print();
 		void createEmptyMatrix();
